Add an a2jmidid-8 ebuild.
I was originally going to take an ebuild off of p.g.o, but I completely reworked
it to use the waf-utils eclass (and the DOCS variable). I also changed the
DEPEND/RDEPEND definitions to match the default used in the gentoo syntax files
for vim, which make more sense than what was done originally anyway.