Add pd-base/pd-extended-0.42.5 ebuild.
This is a modified version of the ebuild available in the overnight overlay,
which is itself derivative of the ebuild in pd-overlay.  This is because the
overnight version is cleaned up a bit and uses a newer EAPI version.  This
version of the ebuild introduces the following changes:

- Move the contents of the $Linux and $externals variables to IUSE_PD_EXTERNALS.
  This is because various externals would be skipped, such as the hid external,
  regardless of USE flag settings.  I wish I knew what those variables were
  supposed to achieve in the first place, they were not used anywhere but IUSE.

Externals that don't build on my system are deactivated by default, feel free to
get them to build.