Add *untested* Giada 0.8.4 ebuild
I wanted to see if this version worked with newer Rtaudio releases, but it
requires RtMidi, which installs RtError.h, which causes a merge conflict with
RtAudio.  The newest version of RtMidi merged RtError.h into RtMidi.h, but
changed the API a bit, so that Giada does not work with it yet.  Therefor, this
ebuild is as yet untested.