Add an ebuild for udiskie 1.0.2
This version of udiskie switched to pygobject-3, and otherwise switched
dependencies around a bit.

In addition, the ebuild needs two additional sed calls to avoid importing the
udiskie module, since if udiskie is already installed, the import uses the
global version (I tried to figure out how to change that and failed, setting
PYTHONPATH didn't work).  This is somewhat OK, since setup.py only uses the
__version__ property, anyway, so the sed replaces that with the value of ${PV}.