Add an ebuild for FAUST 2.0.a30
I had to remove some USE flags because I don't have the time to update my
patches.  Perhaps this will be easier with their new CMake build system?  I
haven't had a detailed look at it, so I don't know whether it's ready or not.

I also bumped the llvm dependency to >=3.5.0.  I don't know for sure if older
versions would also work, but since the 20140317 git snapshot stopped compiling
for me after upgrading to llvm 3.5, I suppose that this version of FAUST will
not compile with versions before that.