Remove pyclewn
Version 2.0.1 has now appeared in the main portage tree.

The problem is that pyclewn 2.x creates its Vim run-time files in the form of a
vimball via a Python call, as opposed to just including them in a source
archive.  This is apparently to make distribution through PIP feasible, which
is understandable.  However, this made packaging the entirety of pyclewn
difficult, since vimballs can't be used as-is.

This is something that my ebuild didn't handle, but that the new ebuild in the
main tree does by using the new package app-arch/vimball.  Note that this also
obsoletes the old 1.11 version.