sci-misc/zotero-bin: fix SRC_URI handling
Fix the handling of SRC_URI in sci-misc/zotero-bin by utilizing USE
conditionals in the SRC_URI string.  This has the advantage that the
package manager now knows of all possible SRC_URIs, and allows pkgcheck
to properly deduce the correct one (for reasons unknown to me it picked
the SRC_URI for x86).

Signed-off-by: Marc Joliet <marcec@gmx.de>